Finished exercies 10.6

+++ app/controllers/users_controller.rb
@@ -1,7 +1,8 @@
 class UsersController < ApplicationController
-  before_filter :authenticate, :only  => [:index, :edit, :update, :destroy]
-  before_filter :correct_user, :only  => [:edit, :update]
-  before_filter :admin_user,   :only => :destroy
+  before_filter :authenticate,  :only  => [:index, :edit, :update, :destroy]
+  before_filter :correct_user,  :only  => [:edit, :update]
+  before_filter :admin_user,    :only  => :destroy
+  before_filter :non_signedin_actions, :only  => [:new, :create] 
   
   def new
     @title = "Sign up"
@@ -48,8 +49,12 @@ class UsersController < ApplicationController
   
   def destroy
     user = User.find(params[:id])
-    user.destroy
-    flash[:success] = "User #{user.name} destroyed."
+    if user.email != current_user.email 
+      user.destroy
+      flash[:success] = "User #{user.name} destroyed."
+    else
+      flash[:error] = "Can't destroy the admin users themselves"
+    end
     redirect_to users_path
   end
   
@@ -66,5 +71,9 @@ class UsersController < ApplicationController
     
     def admin_user
       redirect_to(root_path) unless current_user.admin?
+    end
+    
+    def non_signedin_actions
+      redirect_to(root_path) unless current_user.nil?
     end
 end
